Adopt pyupgrade and targeted RUF lint rules, apply fixes
Add UP, RUF012, RUF021, RUF022, and RUF046 to the ruff gate. Sort the
__all__ lists that had drifted, parenthesize mixed and/or chains, drop
redundant int() casts around round/ceil/floor, annotate mutable class
constants with ClassVar, and move RDKitResult and molfile._chunked to
PEP 695 type-parameter syntax. Test doubles keep mutable class state,
so tests ignore RUF012.

+++ app/ui/export_raster_renderer.py
@@ -48,7 +48,7 @@ def export_raster_file(
     height_px = max(1, round(plan.out_h_pt / POINTS_PER_INCH * dpi))
     with exported_scene(scene, export_items):
         image = QImage(width_px, height_px, QImage.Format.Format_ARGB32_Premultiplied)
-        dots_per_meter = int(round(dpi / METERS_PER_INCH))
+        dots_per_meter = round(dpi / METERS_PER_INCH)
         image.setDotsPerMeterX(dots_per_meter)
         image.setDotsPerMeterY(dots_per_meter)
         image.fill(Qt.GlobalColor.white if background == "white" else Qt.GlobalColor.transparent)
